--- src/types/popularCity.ts
@@ -1,6 +1,0 @@
-export interface PopularCity {
-  cityId: number;
-  cityName: string;
-  hotelCount: number;
-  cityImage: string;
-}
